# api/openapi/order.yml
openapi: 3.0.0
info:
  title: order service
  description: order service
  version: 1.0.0
servers:
  - url: 'https://{hostname}/api'
    variables:
      hostname:
        default: 127.0.0.1
paths:
  /customer/{customerId}/orders/{orderId}:
    get:
      description: "get order"
      parameters:
        - in: path
          name: customerId
          schema:
            type: string
          required: true

        - in: path
          name: orderId
          schema:
              type: string
          required: true

      responses:
        '200':
          description: todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

        default:
          description: todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /customer/{customerId}/orders:
    post:
      description: "create order"
      parameters:
        - in: path
          name: customerId
          schema:
            type: string
          required: true

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'

      responses:
        '200':
          description: todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

        default:
          description: todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: string
          description: order id
        customerId:
          type: string
          description: customer id
        status:
          type: string
          description: order status
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        paymentLink:
          type: string
          description: payment link

    Item:
      type: object
      properties:
        id:
          type: string
          description: item id
        name:
          type: string
          description: item name
        quantity:
          type: integer
          format: int32
          description: item quantity
        priceId:
          type: string
          description: item price id

    Error:
      type: object
      properties:
        message:
          type: string
          description: error message

    CreateOrderRequest:
      type: object
      required:
        - customerId
        - items
      properties:
        customerId:
          type: string
          description: customer id
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemWithQuantity'

    ItemWithQuantity:
        type: object
        properties:
          id:
            type: string
            description: item id
          quantity:
            type: integer
            format: int32
            description: item quantity