# internal/order/ports/openapi_types.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package ports

// CreateOrderRequest defines model for CreateOrderRequest.
type CreateOrderRequest struct {
	// CustomerId customer id
	CustomerId string             `json:"customerId"`
	Items      []ItemWithQuantity `json:"items"`
}

// Error defines model for Error.
type Error struct {
	// Message error message
	Message *string `json:"message,omitempty"`
}

// Item defines model for Item.
type Item struct {
	// Id item id
	Id *string `json:"id,omitempty"`

	// Name item name
	Name *string `json:"name,omitempty"`

	// PriceId item price id
	PriceId *string `json:"priceId,omitempty"`

	// Quantity item quantity
	Quantity *int32 `json:"quantity,omitempty"`
}

// ItemWithQuantity defines model for ItemWithQuantity.
type ItemWithQuantity struct {
	// Id item id
	Id *string `json:"id,omitempty"`

	// Quantity item quantity
	Quantity *int32 `json:"quantity,omitempty"`
}

// Order defines model for Order.
type Order struct {
	// CustomerId customer id
	CustomerId *string `json:"customerId,omitempty"`

	// Id order id
	Id    *string `json:"id,omitempty"`
	Items *[]Item `json:"items,omitempty"`

	// PaymentLink payment link
	PaymentLink *string `json:"paymentLink,omitempty"`

	// Status order status
	Status *string `json:"status,omitempty"`
}

// PostCustomerCustomerIdOrdersJSONRequestBody defines body for PostCustomerCustomerIdOrders for application/json ContentType.
type PostCustomerCustomerIdOrdersJSONRequestBody = CreateOrderRequest